obsel skins: when specifiying an icon source through iconExpression, resize it to fit the Traceline height
If the icon is specified through icon, it is not resized, since it is assumed that the user knows what it is doing, and he may want to do some special presentations (use a small icon for instance).
If the icon is specified through iconExpression, then it probably gets its value from the obsel contents, and we cannot assume that the size will be appropriate. So we force a resize.